Refactor d/service and d/services: split data sources into their own files, add comments, add clearer variable names
